Rename vimiv.vimiv to vimiv.startup
This is what this module does, it starts the main application. Having a
module called vimiv.vimiv is less descriptive. In addition, this allows
executing python scripts that import vimiv from within the vimiv folder
as the vimiv module is no longer shadowed by the local vimiv.py.
