+++ setup.py
@@ -39,7 +39,7 @@ setuptools.setup(
     python_requires=">=3.6",
     packages=setuptools.find_packages(),
     ext_modules=[manipulate_module],
-    entry_points={"gui_scripts": ["vimiv = vimiv.vimiv:main"]},
+    entry_points={"gui_scripts": ["vimiv = vimiv.startup:main"]},
     name="vimiv",
     version=".".join(str(num) for num in read_from_init("version_info")),
     description=read_from_init("description"),
